# Optimize component completion list serialization/deserialization