--- a/src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Completion/OptimizedCompletionList.cs
+++ b/src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Completion/OptimizedCompletionList.cs
@@ -134,7 +134,7 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer.Completion
                     writer.WriteValue(completionItem.InsertText);
                 }
 
-                if (completionItem.InsertTextFormat != default)
+                if (completionItem.InsertTextFormat != InsertTextFormat.PlainText)
                 {
                     writer.WritePropertyName("insertTextFormat");
                     writer.WriteValue(completionItem.InsertTextFormat);