--- a/src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Completion/VSLspCompletionItem.cs
+++ b/src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Completion/VSLspCompletionItem.cs
@@ -1,0 +1,12 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using OmniSharp.Extensions.LanguageServer.Protocol.Models;
+
+namespace Microsoft.AspNetCore.Razor.LanguageServer.Completion
+{
+    internal class VSLspCompletionItem : CompletionItem
+    {
+        public PascalCasedSerializableImageElement Icon { get; set; }
+    }
+}