… src/Razor/benchmarks/Microsoft.AspNetCore.Razor.Performance/taghelpers.json