- Added new benchmarks to capture component completion list serialization/deserialization. To get a proper count for how good/bad it was I updated the taghelpers.json file to be from a project that utilized component vendors (a lotttt more TagHelpers).
- Built a new "OptimizedCompletionList" type that wraps pre-existing completion lists and provides its own `JsonConverter`.
  - The new JsonConverter tries to optimize the serialization speeds of our completion lists by reducing redundant data and caching well-known data.
    - Doesn't duplicate `insertText`, `filterText` and `sortText` if they match the `label` (functions the same).
    - Don't serialize other properties if set to their defaults (`null`, false etc.)
    - Cache the serialization data for our TagHelper icon serialization.
- Refactored `RazorCompletionEndpoint` to be better benchmarkable.

### Baseline:
|                                              Method |     Mean |    Error |   StdDev |  Op/s |     Gen 0 |     Gen 1 |    Gen 2 | Allocated |
|---------------------------------------------------- |---------:|---------:|---------:|------:|----------:|----------:|---------:|----------:|
| 'Component Completion List Roundtrip Serialization' | 60.53 ms | 1.199 ms | 2.282 ms | 16.52 | 2555.5556 | 1111.1111 | 333.3333 |  16.49 MB |
|           'Component Completion List Serialization' | 21.17 ms | 0.416 ms | 0.869 ms | 47.24 |  437.5000 |  312.5000 | 218.7500 |   3.34 MB |
|         'Component Completion List Deserialization' | 39.00 ms | 0.772 ms | 0.858 ms | 25.64 | 2384.6154 | 1076.9231 | 230.7692 |  13.12 MB |

### After
|                                              Method |      Mean |     Error |    StdDev |   Op/s |     Gen 0 |     Gen 1 |    Gen 2 | Allocated |
|---------------------------------------------------- |----------:|----------:|----------:|-------:|----------:|----------:|---------:|----------:|
| 'Component Completion List Roundtrip Serialization' | 32.623 ms | 0.6457 ms | 1.1139 ms |  30.65 | 2125.0000 | 1125.0000 | 187.5000 |  13.66 MB |
|           'Component Completion List Serialization' |  8.734 ms | 0.1084 ms | 0.1014 ms | 114.50 |  312.5000 |  234.3750 | 234.3750 |   2.42 MB |
|         'Component Completion List Deserialization' | 24.501 ms | 0.4178 ms | 0.3908 ms |  40.82 | 1968.7500 |  906.2500 | 125.0000 |  11.21 MB |

**Roundtrip op/s improvement:** 1.85x faster
**Serialization op/s improvement:** 2.42x faster
**Deserialization op/s improvement:** 1.59x faster

Part of dotnet/aspnetcore#29017