@NTaylorMullen 2021-01-05
Oh and the completion list being serialized in the benchmark had 880 items in it. There's still two more optimizations I plan to make to it but going to do those in follow ups