@NTaylorMullen 2021-01-06
Merged in my other [completion list perf PR](https://github.com/dotnet/aspnetcore-tooling/pull/2888) which had additional improvements. New "after" benchmark is:

|                                              Method |      Mean |     Error |    StdDev |   Op/s |     Gen 0 |     Gen 1 |    Gen 2 | Allocated |
|---------------------------------------------------- |----------:|----------:|----------:|-------:|----------:|----------:|---------:|----------:|
| 'Component Completion List Roundtrip Serialization' | 28.129 ms | 0.3938 ms | 0.3684 ms |  35.55 | 2156.2500 | 1187.5000 | 218.7500 |   13.6 MB |
|           'Component Completion List Serialization' |  6.081 ms | 0.0450 ms | 0.0399 ms | 164.46 |  312.5000 |  242.1875 | 226.5625 |    2.4 MB |
|         'Component Completion List Deserialization' | 24.501 ms | 0.4178 ms | 0.3908 ms |  40.82 | 1968.7500 |  906.2500 | 125.0000 |  11.21 MB |

Which results in the following overall improvements
**Roundtrip op/s improvement:** 2.15x faster
**Serialization op/s improvement:** 3.48x faster
**Deserialization op/s improvement:** 1.59x faster